# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for CodeGeeky",
        "title": "CodeGeeky API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/internshipbatches": {
            "get": {
                "description": "List internship batches with optional filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "InternshipBatch"
                ],
                "summary": "List internship batches",
                "parameters": [
                    {
                        "enum": [
                            "upcoming",
                            "ongoing",
                            "completed",
                            "cancelled"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "InternshipBatchStatusUpcoming",
                            "InternshipBatchStatusOngoing",
                            "InternshipBatchStatusCompleted",
                            "InternshipBatchStatusCancelled"
                        ],
                        "description": "These fields are used to filter internships by start and end date",
                        "name": "batch_status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "expand",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "These fields are used to filter internships by internship id",
                        "name": "internship_ids",
                        "in": "query"
                    },
                    {
                        "maximum": 1000,
                        "minimum": 1,
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "These fields are used to filter internships by name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "published",
                            "deleted",
                            "archived",
                            "inactive",
                            "pending"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "StatusPublished",
                            "StatusDeleted",
                            "StatusArchived",
                            "StatusInactive",
                            "StatusPending"
                        ],
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ListInternshipBatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ierr.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new internship batch with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "InternshipBatch"
                ],
                "summary": "Create a new internship batch",
                "parameters": [
                    {
                        "description": "Internship batch details",
                        "name": "batch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateInternshipBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.InternshipBatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ierr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/internshipbatches/{id}": {
            "get": {
                "description": "Get an internship batch by its unique identifier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "InternshipBatch"
                ],
                "summary": "Get an internship batch by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Internship Batch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.InternshipBatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ierr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ierr.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an internship batch by its unique identifier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "InternshipBatch"
                ],
                "summary": "Update an internship batch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Internship Batch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Internship batch details",
                        "name": "batch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateInternshipBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.InternshipBatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ierr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ierr.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an internship batch by its unique identifier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "InternshipBatch"
                ],
                "summary": "Delete an internship batch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Internship Batch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ierr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ierr.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateInternshipBatchRequest": {
            "type": "object",
            "required": [
                "internship_id",
                "name"
            ],
            "properties": {
                "batch_status": {
                    "$ref": "#/definitions/types.InternshipBatchStatus"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "internship_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
        "dto.InternshipBatchResponse": {
            "type": "object",
            "properties": {
                "batch_status": {
                    "$ref": "#/definitions/types.InternshipBatchStatus"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "internship_id": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/types.Metadata"
                },
                "name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/types.Status"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "dto.ListInternshipBatchResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.InternshipBatchResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/types.PaginationResponse"
                }
            }
        },
...
        "dto.UpdateInternshipBatchRequest": {
            "type": "object",
            "properties": {
                "batch_status": {
                    "$ref": "#/definitions/types.InternshipBatchStatus"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
        "ierr.ErrorDetail": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "internal_error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "ierr.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/ierr.ErrorDetail"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "types.InternshipBatchStatus": {
            "type": "string",
            "enum": [
                "upcoming",
                "ongoing",
                "completed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "InternshipBatchStatusUpcoming",
                "InternshipBatchStatusOngoing",
                "InternshipBatchStatusCompleted",
                "InternshipBatchStatusCancelled"
            ]
        },
...
        "types.Metadata": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "types.PaginationResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "types.Status": {
            "type": "string",
            "enum": [
                "published",
                "deleted",
                "archived",
                "inactive",
                "pending"
            ],
            "x-enum-varnames": [
                "StatusPublished",
                "StatusDeleted",
                "StatusArchived",
                "StatusInactive",
                "StatusPending"
            ]
        },
...
    },
    "securityDefinitions": {
        "Authorization": {
            "description": "Enter the token with the `Bearer ` prefix, e.g. `Bearer \u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}